add tests and pass static checking
