+++ gamewinner/strategies/evanmiya/the_cuts.py
@@ -1,5 +1,5 @@
-from typing import no_type_check
 import random
+from typing import no_type_check
 
 from gamewinner.strategies.evanmiya.ievanmiya import IEvanMiyaStrategy
 from gamewinner.team import Team
@@ -11,27 +11,24 @@ class TheCuts23(IEvanMiyaStrategy):
 
     Heavily based on BPR, but with a large randomness factor added as well
 
-    Additional weighting on Resume Rank and a little extra random bonus 
-    weighted on the inverse of Home Rank (i.e. you get a slight penalty 
+    Additional weighting on Resume Rank and a little extra random bonus
+    weighted on the inverse of Home Rank (i.e. you get a slight penalty
     for being significantly better at home than on the road)
     """
 
     @no_type_check
     def _team_metric(self, team: Team) -> float:
         overall_score = (
-            random.random() *
-            self._rank_to_percentile(team.evanmiyaResumeRank) * 
-            team.evanmiyaBPR
+            random.random()
+            * self._rank_to_percentile(team.evanmiyaResumeRank)
+            * team.evanmiyaBPR
         )
 
         # upset factor
-        overall_score = (
-            overall_score + (
-                0.5 * 
-                random.random() * 
-                self._rank_to_percentile(team.evanmiyaHomeRank, reverse = T)
-            )
+        overall_score = overall_score + (
+            0.5
+            * random.random()
+            * self._rank_to_percentile(team.evanmiyaHomeRank, reverse=True)
         )
-        
-        return overall_score
 
+        return overall_score
